pyproject.toml
[project]
name = "cuga"
version = "0.2.14"
description = "CUGA is an open-source generalist agent for the enterprise, supporting complex task execution on web and APIs, OpenAPI/MCP integrations, composable architecture, reasoning modes, and policy-aware features."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10, <3.14"
dependencies = [
    "langchain>=1.0",
    "langchain-ibm>=0.3.9",
    "langchain-openai",
    "langchain-core",
    "aiohttp",
    "httpx",
    "docker",
    "langgraph",
    "markdownify",
    "uvicorn",
    "fastapi[standard]",
    "langfuse",
    "dynaconf",
    "playwright==1.49.0",
    "browsergym-core==0.13.0",
    "python-dotenv",
    "loguru",
    "mcp[cli]>=1.23.0",
    "psutil>=7.0.0",
    "typer>=0.15.3",
    "typer-slim>=0.15.3", # explicit: docling pulls typer meta-package which needs typer-slim
    "fastmcp>=3.2.0",
    "langchain-mcp-adapters",
    "pyyaml>=6.0.3",
    "pymilvus[milvus-lite]>=2.6.4",
    "langchain-groq>=0.3.8",
    "langchain-litellm",
    "litellm>=1.83.0",
    "tavily-python",
    "fastembed>=0.4.0",
    "a2a-sdk[http-server]>=0.3.22",
    "sqlite-vec>=0.1.6",
    "pgvector>=0.4.2",
    "asyncpg>=0.30",
    "PyJWT[crypto]>=2.10.0",
    "cryptography>=44.0.0",
    "hvac>=2.0.0",
    "boto3>=1.34.0",
    "openlit>=1.40.1",
    # Knowledge engine (replaces OpenRAG)
    "langchain-docling>=2.0,<3.0",
    "langchain-ollama>=0.2.0",
    "langchain-text-splitters>=0.3,<0.4",
    "beautifulsoup4>=4.12",
    "psycopg[binary]>=3.2",
]

[project.optional-dependencies]
evaluation = ["rapidfuzz"]
e2b = ["e2b-code-interpreter>=2.4.1"]
memory = ["mem0ai[extras]"]
knowledge = []  # knowledge is now built-in, no extra dependencies needed
health = ["cuga-oak-health; python_version>='3.12'"]

[project.urls]
Homepage = "https://cuga.dev"
Repository = "https://github.com/cuga-project/cuga-agent"

[tool.uv]
package = true
constraint-dependencies = [
    "pillow>=12.1.1",  # CVE-2026-25990: Out-of-bounds write in PSD loading
    "urllib3>=2.6.3",  # CVE-2025-66418, CVE-2025-66471, CVE-2025-50181, CVE-2025-50182, CVE-2026-21441
    "h11>=0.16.0",     # CVE-2025-43859: HTTP request smuggling via chunked encoding
    "setuptools<70",   # milvus-lite requires pkg_resources (removed in setuptools 70+)
]

# openlit 1.40.1 declares openai<2; litellm>=1.83 needs openai 2.x (CVE-2026-35030 / GHSA-jjhc-v7c2-5hh6).
[[tool.uv.dependency-metadata]]
name = "openlit"
version = "1.40.1"
requires-dist = [
    "anthropic>=0.42.0,<1.0.0",
    "boto3>=1.34.0,<2.0.0",
    "botocore>=1.34.0,<2.0.0",
    "openai>=1.1.1,<3.0.0",
    "opentelemetry-api>=1.38.0,<2.0.0",
    "opentelemetry-exporter-otlp>=1.38.0,<2.0.0",
    "opentelemetry-instrumentation>=0.52b0,<1.0.0",
    "opentelemetry-instrumentation-aiohttp-client>=0.52b0,<1.0.0",
    "opentelemetry-instrumentation-asgi>=0.52b0,<1.0.0",
    "opentelemetry-instrumentation-django>=0.52b0,<1.0.0",
    "opentelemetry-instrumentation-falcon>=0.52b0,<1.0.0",
    "opentelemetry-instrumentation-fastapi>=0.52b0,<1.0.0",
    "opentelemetry-instrumentation-flask>=0.52b0,<1.0.0",
    "opentelemetry-instrumentation-httpx>=0.52b0,<1.0.0",
    "opentelemetry-instrumentation-pyramid>=0.52b0,<1.0.0",
    "opentelemetry-instrumentation-requests>=0.52b0,<1.0.0",
    "opentelemetry-instrumentation-starlette>=0.52b0,<1.0.0",
    "opentelemetry-instrumentation-tornado>=0.52b0,<1.0.0",
    "opentelemetry-instrumentation-urllib>=0.52b0,<1.0.0",
    "opentelemetry-instrumentation-urllib3>=0.52b0,<1.0.0",
    "opentelemetry-sdk>=1.38.0,<2.0.0",
    "opentelemetry-semantic-conventions>=0.59b0,<1.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "requests>=2.26.0,<3.0.0",
    "schedule>=1.2.2,<2.0.0",
    "xmltodict>=0.13.0,<1.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
exclude = [
    "frontend_workspaces*"
]

[tool.setuptools.package-data]
cuga = [
    "configurations/**/*.toml",
    "configurations/**/*.md",
    "settings.toml",
    "backend/**/*.jinja2",
    "backend/**/*.js",
    "backend/server/demo_setup_utils/*.json",
    "frontend/dist/**",
]
scripts = ["*.py"]
system_tests = ["**/*.py", "**/*.yaml", "**/*.sh"]



[project.scripts]
load_specs = "scripts.api_utils.load_openapi_specs:main"
petstore = "scripts.commands:run_petstore"
demo = "scripts.commands:run_demo"
registry = "scripts.commands:run_api_registry"
registry_appworld = "scripts.commands:run_api_registry_appworld"

graph = "scripts.draw_graph:main"
eval_api = "scripts.commands:run_eval_api"
appworld_eval = "evaluation.appworld_eval:run_main"
appworld_setup = "scripts.commands:setup_appworld_environment"
appworld_setup_docker = "scripts.commands:setup_appworld_environment_docker"
cuga = "cuga.cli:app"
code = "evaluation.code_generator:main"
exp = "scripts.eval_gui:main"
digital_sales_mcp = "scripts.commands:run_digital_sales_mcp"
digital_sales_openapi = "scripts.commands:run_digital_sales_openapi"
test_llm = "scripts.commands:test_llm"

[dependency-groups]
dev = [
    "pytz>=2025.1",
    "ruff>=0.9.4",
    "tabulate",
    "beartype",
    "nltk>=3.9.3",
    "termcolor",
    "rapidfuzz",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.1.0",
    "cugaviz",
    "pre-commit>=4.3.0",
    "questionary>=2.1.1",
    "rich>=14.2.0",
    "pip-licenses>=5.5.1",
]
google-genai = [
    "langchain-google-genai>=2.0.11",
]
sandbox = [
    "llm-sandbox>=0.1.3",
]

[tool.pytest.ini_options]
anyio_mode = "auto"
markers = [
    "e2e: end-to-end tests requiring live external services (registry, etc.)",
]
